requirements.txt
click==6.7
croniter==0.3.17
dominate==2.3.1
enum-compat==0.0.2
enum34==1.1.6
eventlet==0.21.0
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Cache==0.13.1
greenlet==0.4.12
gunicorn==19.7.1
hiredis==0.2.0
itsdangerous==0.24
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
nose==1.3.7
PyGithub==1.35
PyJWT==1.5.2
python-dateutil==2.6.1
python-dotenv==0.6.5
redis==2.10.5
redisearch==0.6.3
rejson==0.2.1
-e git+git@github.com:RedisLabs/redismodules.com.git@13f6702a07e4b0c3eb1f6b8ef1fc2387cdfa4a27#egg=rmhub
rmtest==0.5.1
rq==0.8.1
rq-scheduler==0.7.0
six==1.10.0
visitor==0.1.3
Werkzeug==0.12.2
